sc_network_sync

Module block_request_handler

Helper for handling (i.e. answering) block requests from a remote peer via the crate::request_responses::RequestResponsesBehaviour.

Structs

BlockRequestHandler

The full block server implementation of BlockServer. It handles the incoming block requests from a remote peer.

FullBlockDownloader

The full block downloader implementation of `BlockDownloader.

Functions

generate_protocol_config

Generates a RequestResponseProtocolConfig for the block request protocol, refusing incoming requests.