Struct sc_network_sync::mock::MockChainSync

pub struct MockChainSync<Block: BlockT> { /* private fields */ }

Implementations

impl<Block: BlockT> MockChainSync<Block>

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl<Block: BlockT> MockChainSync<Block>

pub fn expect_peer_info(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the peer_info method

pub fn expect_status(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the status method

pub fn expect_num_sync_requests(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the num_sync_requests method

pub fn expect_num_downloaded_blocks(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the num_downloaded_blocks method

pub fn expect_num_peers(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the num_peers method

pub fn expect_num_active_peers(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the num_active_peers method

pub fn expect_new_peer(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the new_peer method

pub fn expect_update_chain_info(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the update_chain_info method

pub fn expect_request_justification(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the request_justification method

pub fn expect_clear_justification_requests(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the clear_justification_requests method

pub fn expect_set_sync_fork_request(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the set_sync_fork_request method

pub fn expect_on_block_data(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the on_block_data method

pub fn expect_on_block_justification(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the on_block_justification method

pub fn expect_on_justification_import(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the on_justification_import method

pub fn expect_on_block_finalized(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the on_block_finalized method

pub fn expect_push_block_announce_validation( &mut self ) -> &mut Expectation<Block>

Create an Expectation for mocking the push_block_announce_validation method

pub fn expect_poll_block_announce_validation( &mut self ) -> &mut Expectation<Block>

Create an Expectation for mocking the poll_block_announce_validation method

pub fn expect_peer_disconnected(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the peer_disconnected method

pub fn expect_metrics(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the metrics method

pub fn expect_block_response_into_blocks(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the block_response_into_blocks method

pub fn expect_poll(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the poll method

pub fn expect_send_block_request(&mut self) -> &mut Expectation<Block>

Create an Expectation for mocking the send_block_request method