Struct sc_network_sync::schema::v1::BlockData

pub struct BlockData {
    pub hash: Vec<u8>,
    pub header: Vec<u8>,
    pub body: Vec<Vec<u8>>,
    pub receipt: Vec<u8>,
    pub message_queue: Vec<u8>,
    pub justification: Vec<u8>,
    pub is_empty_justification: bool,
    pub justifications: Vec<u8>,
    pub indexed_body: Vec<Vec<u8>>,
}

Block data sent in the response.

Fields
hash: Vec<u8>

Block header hash.

header: Vec<u8>

Block header if requested.

optional

body: Vec<Vec<u8>>

Block body if requested.

optional

receipt: Vec<u8>

Block receipt if requested.

optional

message_queue: Vec<u8>

Block message queue if requested.

optional

justification: Vec<u8>

Justification if requested.

optional

is_empty_justification: bool

True if justification should be treated as present but empty. This hack is unfortunately necessary because shortcomings in the protobuf format otherwise doesn’t make in possible to differentiate between a lack of justification and an empty justification.

optional, false if absent

justifications: Vec<u8>

Justifications if requested. Unlike the field for a single justification, this field does not required an associated boolean to differentiate between the lack of justifications and empty justification(s). This is because empty justifications, like all justifications, are paired with a non-empty consensus engine ID.

optional

indexed_body: Vec<Vec<u8>>

Indexed block body if requestd.

optional