Struct sc_network_sync::service::mock::MockNetwork

pub struct MockNetwork { /* private fields */ }

Implementations

impl MockNetwork

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockNetwork

pub fn expect_set_authorized_peers(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_authorized_peers method

pub fn expect_set_authorized_only(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_authorized_only method

pub fn expect_add_known_address(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_known_address method

pub fn expect_report_peer(&mut self) -> &mut Expectation

Create an Expectation for mocking the report_peer method

pub fn expect_disconnect_peer(&mut self) -> &mut Expectation

Create an Expectation for mocking the disconnect_peer method

pub fn expect_accept_unreserved_peers(&mut self) -> &mut Expectation

Create an Expectation for mocking the accept_unreserved_peers method

pub fn expect_deny_unreserved_peers(&mut self) -> &mut Expectation

Create an Expectation for mocking the deny_unreserved_peers method

pub fn expect_add_reserved_peer(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_reserved_peer method

pub fn expect_remove_reserved_peer(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_reserved_peer method

pub fn expect_set_reserved_peers(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_reserved_peers method

pub fn expect_add_peers_to_reserved_set(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_peers_to_reserved_set method

pub fn expect_remove_peers_from_reserved_set(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_peers_from_reserved_set method

pub fn expect_add_to_peers_set(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_to_peers_set method

pub fn expect_remove_from_peers_set(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_from_peers_set method

pub fn expect_sync_num_connected(&mut self) -> &mut Expectation

Create an Expectation for mocking the sync_num_connected method