sc_network_sync

Module strategy

SyncingStrategy defines an interface crate::engine::SyncingEngine uses as a specific syncing algorithm.

A few different strategies are provided by Substrate out of the box with custom strategies possible too.

Modules

chain_sync

Contains the state of the chain synchronization process

polkadot

PolkadotSyncingStrategy is a proxy between crate::engine::SyncingEngine and specific syncing algorithms.

state

State sync strategy.

state_sync

State sync support.

warp

Warp syncing strategy. Bootstraps chain by downloading warp proofs and state.

Structs

StrategyKey

The key identifying a specific strategy for responses routing.

Enums

SyncingAction

Traits

SyncingStrategy

Syncing strategy for syncing engine to use