sc_network_sync::strategy::chain_sync

Function validate_blocks

pub fn validate_blocks<Block: BlockT>(
    blocks: &Vec<BlockData<Block>>,
    peer_id: &PeerId,
    request: Option<BlockRequest<Block>>,
) -> Result<Option<NumberFor<Block>>, BadPeer>

Validate that the given blocks are correct. Returns the number of the first block in the sequence.

It is expected that blocks are in ascending order.