sc_network_sync
Module types

Common syncing types.

Structs

BadPeer

A peer did not behave as expected and should be reported.

ExtendedPeerInfo

Info about a peer’s known state (both full and light).

OpaqueStateRequest

Wrapper for implementation-specific state request.

OpaqueStateResponse

Wrapper for implementation-specific state response.

PeerInfo

The sync status of a peer we are trying to sync with

SyncStatus

Syncing status and statistics.

Enums

PeerRequest

PeerRequestType

SyncEvent

Syncing-related events that other protocols can subscribe to.

SyncState

Reported sync state.

Traits

SyncEventStream

SyncStatusProvider

Provides high-level status of syncing.