sc_network_sync

Module warp_request_handler

Helper for handling (i.e. answering) grandpa warp sync requests from a remote peer.

Structs

RequestHandler

Handler for incoming grandpa warp sync requests from a remote peer.

Functions

generate_request_response_config

Generates a RequestResponseProtocolConfig for the grandpa warp sync request protocol, refusing incoming requests.