Crate sc_network_sync

Blockchain syncing implementation in Substrate.

Re-exports

pub use service::syncing_service::SyncingService;

pub use strategy::warp::WarpSyncConfig;

pub use strategy::warp::WarpSyncPhase;

pub use strategy::warp::WarpSyncProgress;

pub use types::SyncEvent;

pub use types::SyncEventStream;

pub use types::SyncState;

pub use types::SyncStatus;

pub use types::SyncStatusProvider;

Modules

block_relay_protocol

Block relay protocol related definitions.

block_request_handler

Helper for handling (i.e. answering) block requests from a remote peer via the crate::request_responses::RequestResponsesBehaviour.

blocks

engine

SyncingEngine is the actor responsible for syncing Substrate chain to tip and keep the blockchain up to date with network updates.

mock

Contains mock implementations of ChainSync and ‘BlockDownloader’.

service

SyncingEngine-related service code

state_request_handler

Helper for handling (i.e. answering) state requests from a remote peer via the crate::request_responses::RequestResponsesBehaviour.

strategy

PolkadotSyncingStrategy is a proxy between crate::engine::SyncingEngine and specific syncing algorithms.

types

Common syncing types.

warp_request_handler

Helper for handling (i.e. answering) grandpa warp sync requests from a remote peer.