Crate sc_network_sync

Blockchain syncing implementation in Substrate.

Re-exports

pub use service::syncing_service::SyncingService;

pub use strategy::warp::WarpSyncConfig;

pub use strategy::warp::WarpSyncPhase;

pub use strategy::warp::WarpSyncProgress;

pub use types::SyncEvent;

pub use types::SyncEventStream;

pub use types::SyncState;

pub use types::SyncStatus;

pub use types::SyncStatusProvider;

Modules

block_relay_protocol

Block relay protocol related definitions.

block_request

Nested message and enum types in BlockRequest.

block_request_handler

Helper for handling (i.e. answering) block requests from a remote peer via the crate::request_responses::RequestResponsesBehaviour.

blocks

engine

SyncingEngine is the actor responsible for syncing Substrate chain to tip and keep the blockchain up to date with network updates.

mock

Contains mock implementations of ChainSync and ‘BlockDownloader’.

service

SyncingEngine-related service code

state_request_handler

Helper for handling (i.e. answering) state requests from a remote peer via the crate::request_responses::RequestResponsesBehaviour.

strategy

SyncingStrategy defines an interface crate::engine::SyncingEngine uses as a specific syncing algorithm.

types

Common syncing types.

warp_request_handler

Helper for handling (i.e. answering) grandpa warp sync requests from a remote peer.

Structs

BlockData

Block data sent in the response.

BlockRequest

Request block data from a peer.

BlockResponse

Response to BlockRequest

KeyValueStateEntry

A key value state.

StateEntry

A key-value pair.

StateRequest

Request storage data from a peer.

StateResponse

Enums

Direction

Block enumeration direction.