Crate sc_network_sync

Contains the state of the chain synchronization process

At any given point in time, a running node tries as much as possible to be at the head of the chain. This module handles the logic of which blocks to request from remotes, and processing responses. It yields blocks to check and potentially move to the database.

Usage

The ChainSync struct maintains the state of the block requests. Whenever something happens on the network, or whenever a block has been successfully verified, call the appropriate method in order to update it.

Modules

block_request_handler

Helper for handling (i.e. answering) block requests from a remote peer via the crate::request_responses::RequestResponsesBehaviour.

blocks

message

Network packet message types. These get serialized and put into the lower level protocol payload.

schema

Include sources generated from protobuf definitions.

state

State sync support.

state_request_handler

Helper for handling (i.e. answering) state requests from a remote peer via the crate::request_responses::RequestResponsesBehaviour.

warp

Warp sync support.

warp_request_handler

Helper for handling (i.e. answering) grandpa warp sync requests from a remote peer.

Structs

BadPeer

A peer did not behave as expected and should be reported.

ChainSync

The main data structure which contains all the state for a chains active syncing strategy.

Metrics

PeerInfo

The sync status of a peer we are trying to sync with

PeerSync

All the data we have about a Peer that we are trying to sync with

Status

Syncing status and statistics.

Enums

AncestorSearchState

The ancestor search state expresses which algorithm, and its stateful parameters, we are using to try to find an ancestor block

OnBlockData

Result of ChainSync::on_block_data.

OnBlockJustification

Result of ChainSync::on_block_justification.

OnStateData

Result of ChainSync::on_state_data.

PeerSyncState

The state of syncing between a Peer and ourselves.

PollBlockAnnounceValidation

Result of ChainSync::poll_block_announce_validation.

SyncMode

Operation mode.

SyncState

Reported sync state.